changed: put musicvideo videodatabase methods in separate file
